changed elapsed time to a long.

+++ java-testing-framework/src/main/java/com/enterprise/model/Stopwatch.java
@@ -22,7 +22,7 @@ public class Stopwatch {
         elapsedTICKTOCK = startedTICK.until(endedTOCK, NANOS);
     }
 
-    public double getElapsedTime()
+    public long getElapsedTime()
     {
         return elapsedTICKTOCK;
     }
